+++ golem-worker-executor-base/tests/guest_languages.rs
@@ -525,31 +525,3 @@ async fn c_example_2() {
 
     check!(first_line == format!("Hello C! {year}"));
 }
-
-#[tokio::test]
-#[tracing::instrument]
-async fn swift_example_1() {
-    let context = TestContext::new();
-    let executor = start(&context).await.unwrap();
-
-    let component_id = executor.store_component("swift-1").await;
-    let worker_id = executor.start_worker(&component_id, "swift-1").await;
-
-    let mut rx = executor.capture_output(&worker_id).await;
-
-    let _ = executor
-        .invoke_and_await(&worker_id, "wasi:cli/run@0.2.0/run", vec![])
-        .await
-        .unwrap();
-
-    tokio::time::sleep(Duration::from_secs(5)).await;
-    let lines = events_to_lines(&mut rx).await;
-
-    drop(executor);
-
-    let now = chrono::Local::now();
-    let year = now.year();
-
-    check!(lines[0] == "Hello world!".to_string());
-    check!(lines[1] == year.to_string());
-}
